some progress on refactoring
